EPMRPP-104683: moved EXECUTION_STATUSES to a shared place
