+++ app/src/pages/inside/manualLaunchesPage/manualLaunchDetailsPage/manualLaunchExecutions/types.ts
@@ -22,10 +22,3 @@ export interface ManualLaunchExecutionsProps {
   pageInfo: Page | null;
   isLoading: boolean;
 }
-
-export enum EXECUTION_STATUSES {
-  FAILED = 'FAILED',
-  PASSED = 'PASSED',
-  SKIPPED = 'SKIPPED',
-  TO_RUN = 'TO_RUN',
-}
