+++ app/src/pages/inside/manualLaunchesPage/manualLaunchExecutionPage/executionStatusButtons/executionStatusButtons.tsx
@@ -17,11 +17,11 @@
 import { useIntl } from 'react-intl';
 
 import { createClassnames } from 'common/utils';
+import { EXECUTION_STATUSES } from 'controllers/manualLaunch';
 
 import { STATUS_BUTTONS } from '../constants';
 import type { ExecutionStatusButtonsProps, } from '../types';
 import { useExecutionStatusModal } from '../executionStatusConfirmModal';
-import { EXECUTION_STATUSES } from '../../manualLaunchDetailsPage/manualLaunchExecutions/types';
 
 import styles from './executionStatusButtons.scss';
