+++ app/src/pages/inside/manualLaunchesPage/manualLaunchExecutionPage/executionStatusPopover/executionStatusPopover.tsx
@@ -19,13 +19,13 @@ import { useIntl } from 'react-intl';
 import { Popover } from '@reportportal/ui-kit';
 
 import { createClassnames } from 'common/utils';
+import { EXECUTION_STATUSES } from 'controllers/manualLaunch';
 import { Divider } from 'pages/inside/projectSettingsPageContainer/content/elements';
 
 import { STATUS_CONFIG } from '../constants';
 import { useExecutionStatusModal } from '../executionStatusConfirmModal';
 import { messages } from './messages';
 import { messages as commonExecutionMessages } from '../messages';
-import { EXECUTION_STATUSES } from '../../manualLaunchDetailsPage/manualLaunchExecutions/types';
 
 import styles from './executionStatusPopover.scss';
